Scripting

+++ L3gion/src/Platform/OpenGL/OpenGLShader.cpp
@@ -360,9 +360,9 @@ namespace L3gion{
 		for (const auto& resource : resources.uniform_buffers)
 		{
 			const auto& bufferType = compiler.get_type(resource.base_type_id);
-			uint32_t bufferSize = compiler.get_declared_struct_size(bufferType);
+			size_t bufferSize = compiler.get_declared_struct_size(bufferType);
 			uint32_t binding = compiler.get_decoration(resource.id, spv::DecorationBinding);
-			int memberCount = bufferType.member_types.size();
+			size_t memberCount = bufferType.member_types.size();
 
 			LG_CORE_TRACE("  {0}", resource.name);
 			LG_CORE_TRACE("    Size = {0}", bufferSize);
